requirements.txt
streamlit
openai
pydantic
pymupdf
pillow
numpy
python-dotenv